Update Readme, Rename package, Relative Import
