+++ README.md
@@ -1,16 +1,19 @@
-# uz_inet_traffic_calc
+# Uznet uchun Traffic kalkulator - Traffic Calc 
 [![Build Status](https://travis-ci.org/mirzaevaziz/uz_inet_traffic_calc.svg?branch=master)](https://travis-ci.org/mirzaevaziz/uz_inet_traffic_calc)
 
 Если Вы используете интернет и у Вас не Unlim план, то этот скрипт поможет Вам посчитать сколько траффика в день Вы можете истратить.
 
+
+Agar siz internet foydalanuvchisi bo'lsangiz va sizda Cheksiz(Unlim) tarifi bo'lmasa, mazkur skript kunlik ishlatilgan MB miqdori hamda yana kun davomida qancha ishlatishingiz mumkinligini ko'rsatadi
 ![alt tag](https://github.com/mirzaevaziz/uz_inet_traffic_calc/blob/master/Sarkor_calc_screen.png?raw=true)
 
-##### How to install
+##### Qanday o'rnatish
     $ pip install -r requirements.txt 
     
-##### Install with virtualenv
+##### Virtualenv yordamida o'rnatish 
     $ mkvirtualenv env
-##### Dependencies
+    $ pip install -r requirements.txt 
+##### Bog'liq kutubxonalar (Dependencies)
     * Python 3.x
     * Robobrowser
     * Tabulate
